Variable Declaration Generator
This commit brings to the software the ability to generate assembly code regarding variable declarations inside the scope of a class.
Right now the types available for generation are int, String and boolean.
The right side of an int or boolean declaration can contain arithmetic expressions, for the former, and relational expressions for the latter.
This update aso introduces the method geraAssemblyVariableDeclarator which can be used to generate declarations in other scopes.